+++ src/briefy/reflex/__init__.py
@@ -1,6 +1,9 @@
 """briefy.reflex."""
-import logging
+from zope.configuration.xmlconfig import XMLConfig
 
+import logging
+import briefy.common
+import briefy.gdrive
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -8,3 +11,7 @@ logger.setLevel(logging.INFO)
 cs = logging.StreamHandler()
 cs.setLevel(logging.INFO)
 logger.addHandler(cs)
+
+
+XMLConfig('configure.zcml', briefy.common)()
+XMLConfig('configure.zcml', briefy.gdrive)()