+++ HISTORY.rst
@@ -5,4 +5,6 @@ History
 0.1.0 (Unreleased)
 ------------------
 
-* TODO
+    * Celery code to define app, config, worker and tasks module (rudaporto).
+    * Tasks module to query orders accepted on Leica (rudaporto).
+    * Tasks module to query in google drive folder contents of delivery folder (rudaporto).